Added statistic enrichment for all bankconnections

+++ ebics-dbmodel/src/main/kotlin/org/ebics/client/api/trace/TraceStatistic.kt
@@ -2,5 +2,5 @@ package org.ebics.client.api.trace
 
 interface TraceStatistic {
     val bankConnectionId: Long
-    val traceEntryCount: Long
+    val traceEntryCount: Int
 }
